Refactor: 친구 코드 등록 모달을 컴포넌트로 뺌, 재사용
